pyproject.toml
[project]
name = "really-deep-research"
version = "0.1.0"
description = "Deep Research = make it deep + feed it clean + simple orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "litellm>=1.69.3",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

